Configure: Updated Python3 checks for venv and different versions
Until now, the Python interface was built conditionally if Python3 or
greater, with NumPy installed on it was found. The checks performed for
this in the configure.ac script used both 'AC_CHECK_PROG' and
'AM_PATH_PYTHON' macros which didn't account for some special cases as
mentioned below.

1. AC_CHECK_PROG checked for the 'python3' command, this meant that Python
   would not have been detected for systems that had multiple python3
   versions or just a different name for their python3 interpreter (for
   example 'python3.8'). AM_PATH_PYTHON macro solves this issue, by
   checking for python versions greater than or equal to a certain version
   (3.0 in our case). However, the further checks for NumPy were also
   dependent on the the success of 'AC_CHECK_PROG'.

2. For virtual environements, in some cases 'AM_PATH_PYTHON' would detect
   the global Python Interpreter instead of the virtual env. This meant
   that if a user had Python3 and NumPy installed in their virtual env. it
   would not be detected.

3. The '--without-python' option was not documented in the book and Python
   was not mentioned as an optional dependency.

With this commit, both the issues above have been addressed: 1)
'AC_CHECK_PROG' has been removed. Therefore, only 'AM_PATH_PYTHON' is used,
which is more general. 2) It is now possible to set the 'PYTHON'
environment variable to point to the virtual env's Python before running
the configure script. A cartouche has been added in the optional
dependencies section of the book regarding this. 3) The '--without-python'
option is now documented in the configuration options and Python has been
added as an optional dependency.