Configure: warning is activated if Numpy+Python3 not found
Until now, the 'anywarnings' variable wasn't set when Numpy+Python3
couldn't be found. Also, when '--without-python' was called, the 'no' from
the output of the check for Numpy would still be printed!

With this commit, 'anywarnings' has been added after the Numpy check,
allowing warnings to be printed if they aren't found. Also, the printing of
'yes' or 'no' is printed when the check for Numpy is actually done.