Make (text-not-contains): new Make extension function
Until now, we only had an operator to select the words that contain a
certain string. Similarly, it also regularly happens that you want words
that don't contain a certain string.

With this command, this has been done with the newly added
'ast-text-not-contains' function. Since the internal structure of this
function was almost identical to the 'ast-text-contains' function, a "base"
function has been defined and both call that (with an argument to specify
the differing condition).

Also, while looking over the PDF book, I noticed two places that the
example code had passed the page margin! So they have also been corrected
in this commit.