Make (version-is): new function to help in reproducibility
Until now, to check the version of the running Gnuastro in a Makefile (to
ensure reproducibility), it was necessary to run a program with the
'--version' function and extract the version string from that using Make's
'$(shell ...)' function. This was dirty and slow.

With this commit, a new Gnuastro Make extension function has been added for
this purpose: '$(ast-version-is STRING)'. This enables easy checking of the
version string in a Makefile, as demonstrated in the minimal working
example of the manual.