Book: corrected typo in example of ast-version-is Make function
Until now (in the previous commit), the minimal working example wouldn't
show the special '$@' variable of Make correctly (it was shown as '$')!
This was because '@' is a special character in Texinfo and needs to be
commented if it is to be used in the output.

With this commit, in the Texinfo source, it is now written as '$@@', thus
producing the correct '$@' in the finally produced manual.