Library (statistics.h): std-dev calculation puts value in double
Until now, when calculating the standard deviation, to find the sum of the
squares, we would simply multiply the values in their native data type,
then put the value in a 'double' type. However, this would cause integer
overflow when the values were larger than the square root of the maximum
value! Therefore a next sanity check would produce a zero standard
deviation.

With this commit, to avoid such situations, a temporary variable of type
'double' has been introduced in the two functions that do standard
deviation parsing. The raw value (in any type) is first written into this
variable, and then used in the sum and sum of squares. This fixes the
problem.

This fixes bug #63013.