psf-scale-factor: optionally saving result in output file
Until now, this script was not saving the output result into the output
file, even when it was provided with the option '--output'. This was simply
because a final step to write the result into the output file was missing,
and it was always being printed on standard output.

With this commit, this bug has been fixed. If the user provides a filename
to '--output' (in which the result should be kept), the scale factor number
will be written in that output file (in plain-text).

This fixes bug #63022.