Library (config.h): added macros for python and gnumake checks
Until now, the recently added checks for the availability of Python or GNU
Make's extension libraries weren't provided to the user of the library
through macros.

With this commit, two new macros have been defined in 'gnuastro/config.h'
for informing the library user if these are available or not.

I also arranged the macros in 'lib/gnuastro-internal/config.h.in' (which
acts as the basis for the built 'config.h', in order of increasing macro
name length.