psf-scale-factor: sub-pixel centering while computing the PSF scale factor
Until now, no sub-pixel centering of the input star was done for computing
the scaling factor of the PSF. But this could cause variations in the
scaling of the PSF because of the non perfect aligntment of the PSF and the
target star.

With this commit the sub-pixel centering procedure has been added. Now the
computation of the scaling factor is done on a sub-pixel centered star
stamp. In addition to that, a small bug has been corrected in the
'psf-subtract' script. It was a systematic subraction of '1' in the
shifting of the PSF while positioning it in the place of the star.