Book: correcting '--widthinpix' info and values in PSF scripts sections
Until now, in the Gnuastro book, the description of the option
'--widthinpix' was not properly updated after it was changed in the PSF
scripts. The main change was that it now consider two values for accounting
for different sizes along x and y.

With this commit, the description and values along the documentation book
has been changed. Now it shows the proper options and values.

Issue reported by Giulia Golini.