Library (dimension.h): variable initialization for compiler warnings
Until now, the newly added sort-based collapsing functions in the library
would cause compilation warnings about un-initialized values (which were
logically no problem, but the compiler can't know). This also happend in
the function of 'bin/arithmetic/arithmetic.c' that called those library
functions.

With this commit, the problematic variables are initialized with basic
values that will cause a crash when not set later (and help find/fix the
bug in the unexpected case that it happens!).