Library (python): Fixed import_array warnings
Until now, when trying to build the Python Interface files(`lib/python.c`)
the NumPy C-API would throw "defined but not used" warnings for the
`import_array()` function. In order to import the NumPy C-API in Python
Extension modules we need to call `import_array()` in the module
initialization. However, since the Python Interface only provides utility
functions for the actual Python Extension modules which are defined in
the pyGnuastro(Gnuastro Python package) source, `import_array()` need/
can not be called here.

With this commit, we define the `NO_IMPORT_ARRAY` macro before
including the NumPy C-API headers. This indicates, that the `python.c`
file has no extension modules, and thus no calls to `import_array()`.

Also, fixed a typo in the `gal_fits_img_write` definition in the book.