Warp: aligning the pixel and WCS coords, while correcting distortions
Until now, Warp could only do linear resampling of the input grid. However,
astronomical imaging cameras often have distortions that also need to be
corrected prior to stacking of the images into one pixel grid. Furthermore
the '--align' operator (which tried to estimate the necessary linear
rotation on images with no distortion) would also fail in some peculiar
scenarios.

With this commit, Warp will now construct the output pixel grid and WCS
based on the user desired properties and correct any form of distortion in
the process. This is now the default operation of Warp (when no linear
operations are requested). Until now, Warp would crash with an error if no
linear operations were given.

Unlike Warp's previous linear warps, the new distortion correcting aligning
operations of warp have been written in a newly added library in Gnuastro:
'warp.h'. The Warp program calls these library functions for this job. But
as a library, it can be used in many other scenarios also, not just the
Warp program.

Furthermore, given that the Warp program was one of the first written
programs in Gnuastro, some aspects of it weren't "modern" (for example it
didn't use the high-level 'gnuastro/threads.h' library features and had a
low-level implementation of it, or it didn't follow the more recent naming
conventions in its function names). Such things have also been corrected.

The following changes have also been made with this commit:

 - The linear '--align' option has been removed: this is now the default
   behavior of Warp.

 - Some memory leaks have been found and corrected.