psf-scale-factor: missing 'quiet' option to Crop and Warp calls added
Until this commit, there were some calls to 'astcrop' and 'astwarp' without
the option '$quiet'. As a consequence, even when the user asked for
'--quiet' mode from  the command-line, the script printed out more lines
than just the result.

With this commit, the quiet option ('$quiet') has been added to these
invokations of 'astcrop' and 'astwarp'. In addition to that, several minor
changes have been done with the aim of having proper identations and the
full '--quiet' option instead of '-q'.