Book: clarifying how to deal with no-strings in figures annotation
Until this commit, information on how to not print text for some markers
was not present in the book. Also, how to have different text lengths into
the same table was missing.

With this commit, several phrases have been added in order to clarify how
to not print text on figures. This is important for the user when having
some markers with text and other without text. In addition to that, a
couple of typos have been also corrected.