Library (error.h): Added a Error Handling module
Until now, all of Gnuastro's library functions  performed error
handling using 'error(EXIT_FAILURE, ....)'; thus exiting the program
whenever and error was encountered with a detailed error message.
This wasn't a probelm for the Gnuastro programs however for other
callers like pyGnuastro, this is problematic as it exits from the entire
Python environment.

With this commit, we add a new module to the library `error.h` for
error handling. A new `gal_error_t` structure is defined with a error
typecode, a backend and frontend error message string, a is_warning
flag and next pointer.

The aim is for every function in the library to have an extra error argument
that will hold the error structure. New error messages will be added to the
list.