Library (error.h): Added a Error Handling module
Until now, all of Gnuastro's library functions  performed error
handling using 'error(EXIT_FAILURE, ....)'; thus exiting the program
whenever and error was encountered with a detailed error message.
This wasn't a probelm for the Gnuastro programs however for other
callers like pyGnuastro, this is problematic as it exits from the entire
Python environment.

With this commit, we add a new module `error.h` to the library for
error handling. As part of this module:
- A new `gal_error_t` structure which acts as a linked list of errors.
  It is defined with a:
	-- 8 bit library code(`lib_code`) indicating which library
	   generated the error. All libraries have their own unique
	   code defined in alphabetical order inside an enum.
	-- 8 bit error typecode(`code`) indicating the type of error within
	   that library. Each library has it's own error types defined as
	   macros.
	-- 8 bit flag indicating if the error is a breaking error or just a
	   warning(`is_warning`).
	-- a backend error message string(`back_msg`) which will be generated
	   by the library. Since rthe library already had well defined
	   messages for the different errors, they are kept the same.
	-- a frontend error message string(`front_msg`) which can be added by
	   the user.
- Functions have been defined in the module to:
	-- allocate a new error structure
	-- add a back_msg, front_msg to an error
	-- print all the errors
	-- check if a breaking error occurred with the last library call
	-- check if an error of a particular type exists in the structure

The aim is for every function in the gnuastro library to have an extra
error argument(as their last argument) that will hold the error structure.
New error messages will be pushed onto the stack of errors. And the user
can parse the structure using the functions provided.

Additionally, we define error codes as macros for the different types of
errors in the cosmology library. The values of the macros are set using a
`BITSET` macro function in the error module that returns the `macro_val`
given the `lib_code`, `code` and `is_warning` flag. So, the 32 bit error
macro values, will have the least significant 8 bits(bit 0 - 7) represent
the `is_warning` flag, the next significant 8 bits(bit 8 - 15) represent
the `code` and the next 8 bits(bit 16 - 23) represent the `lib_code`.

Also until now, the error/sanity checks for the binary programs
were performed in their `ui.c` files. Before passing the values
to the actual code of the program. With the addition of the `error.h`
module, we can fortify this system further by adding verfication checks
in the main programs too, incase any errors were missed by the ui sanity
checks.

Also, added the additional error argument required by
the cosmology programs to the library calls made in `cosmiccal.c`.