Book: added new section in wcs.h description on freeing
Until now, the example on how to properly free a WCS structure was only in
the description of the 'gal_wcs_create' function. However, there were two
issues: 1) It wasn't a complete program for someone to just copy-paste. 2)
It can be useful for other functions that read a WCS structure also.

With this commit, the example has been completed and moved to a whole new
section under the description of wcs.h functions.