.mailmap: corrected the email address of Elham Saremi
Until now there were two names and emails for Elham Saremi in the Git
commit history. This caused a repeat in the author list that is printed in
the second page of the book.

With this commit, Git's '.mailmap' file has been edited so when generating
the author list, Git uses the same name and email in both scenarios. This
was confirmed with Elham Saremi in a private communication.