Warp: new --gridfile option to warp one image to grid of another
Until now, when you already had one image with a certain WCS, it was
annoying (when it was aligned: using the three '--center', '--cdelt' and
'--widthinpix') or impossible (when it wasn't aligned and had distortions)
to map another image to that grid using Warp.

With this commit, a new option has been added for this job: '--gridfile'
(and a corresponding '--gridhdu'). When this is given, the WCS and pixel
grid of the output will exactly match that to of the input. This has been
implemented in a library level, so it was necessary to make some changes
there also.

In the process, it was also recognized that there were several cases where
'etc.' came without the last 'period'. Also, since 'etc.' stands for 'et
cetera' which means 'and the rest', using 'and' before 'etc.' is
redundant. One last case, is where there is a 'for example' or 'like' or
'such as' in the sentence; Using 'etc.' while counting examples or cases is
incorrect (see [1] and [2]). For example:

(INCORRECT) Fruits such as apples, oranges, strawberries, etc.
(CORRECT)   Fruits such as apples, oranges, and strawberries.
(CORRECT)   Apples, oranges, strawberries, etc.

References:
[1] https://www.grammarly.com/blog/et-cetera-etc/
[2] https://www.wikihow.com/Use-%22Etc.%22-Correctly