Warp: fix bug when using 'center' truncated image
Until now, Warp used only the outer edges of the input image to measure the
output image size automatically, and did not take the '--center' into
account. For this reason, when a '--center' was defined, Warp would not
calculate the size differently. In this case, size remains the same, but
center is different; which leads to a truncated output image.

With this commit, Warp calculates the output image size based on the given
'--center' value. It will measure the difference between the given
'--center' and the edges of the input image and pick the maximum
value. This value is only half of the image size (i.e. distance from center
to the furthest edge of the image), so it must be multiplied by two
afterwards.

This new method fixed this issue.