NEWS and README: updates to fit new Warp features
Until now, Pedram wasn't acknowedged as the author of the newly added
features for Warp in the NEWS file. Also, the README file's description of
the Warp program hadn't been updated.

With this commit, Pedram has been mentioned as the author of those newly
added features in Warp. I also noticed that in the README file we still had
the old name of 'astfits', which has now been corrected. Furthermore, I
noticed some typos in two error messages that have been corrected.