Book: better example for the usage of --asis
Until now, the example usage of the '--asis' operator of the 'astfits'
program was to select keys from one file and write them to another. This
was written in a time when the '--copykeys' option's keyword name selection
method hadn't yet been implemented! Since '--copykeys' now does the job
easily, the explanation was not useful and would just confuse a reader!

With this commit, a new use case has been added that was also one of the
original motivations behind this option: to add FITS keywords from a
new-line separated plain-text file.

This example was added after a discussion with Giulia Golini and Raul
Infante-Sainz.