Book: added hashbang in the script of the Sufi simulation tutorial
Until now, the details of writing a script weren't mentioned before the
Sufi simulation tutorial. Therefore the example script at the end of that
tutorial didn't have a hashband. After displacing the tutorials, the Sufi
tutorial has come to the end. By this time, the reader has learned every
detail about a script at the end of the general tutorial.  Therefore we can
include details of a script's body!

With this commit, the hashband "#!/bin/bash" has been added in the first
line of the script to say the shell program to use is GNU Bash. Also a link
to the script writing section of the general tutorial has also been added.