Book: updated the section in ConvertType for aligning inputs
Until now, there was a section in ConvertType which used a very crude way
for aligning color channels with offsets before putting them in one color
image: it used Crop, and wouldn't correct for distortions.

With this commit, using the new WCS aligning feature of Warp, that section
has been almost fully re-written (and greatly simplified).