Warp: implemented --width and --widthinpix options like Crop
Until now, in Warp, we only had the '--widthinpix' option which would
accept values in pixels. However, in some cases, the users may want to
specify the width in units of degrees, arcminutes or arcseconds. Also,
because Gnuastro users are already use to Crop's definitions of '--width',
this behavior of Warp was confusing.

With this commit, Warp now accepts the main width of the output through the
'--width' option (similar to Crop), but also accepts '--widthinpix' for
cases when the user wants to give the width in pixels (again, similar to
Crop).

This was suggested by Raul Infante-Sainz.