Statistics: test data for checking fitting in make check
Until now, the test file to test the newly added least squares fitting
features of Statistics wasn't available in the tarball (not defined in
'EXTRA_DIST' of 'tests/Makefile.am'). As a result, the test would be
skipped.

With this commit, it has been given as a value to that variable, thus
ensuring that it will be placed in the tarball upon creating a
distribution.

I also noticed (while testing a clean build), that the Gnulib 'strcasecmp'
module has been removed and the respective function is now available in the
'strcase' module (which was already included in 'bootstrap.conf'). So with
this commit, 'strcasecmp' has also been removed from there.