Warp: added Pedram as author and citation to paper in preparation
Until now, the Warp program's only authors was Mohammad, but with Pedram
has contributed significantly to the current version of the program. It was
therefore necessary to add him as an author. Pedram is also working on a
paper to describe the methodology of Warp, and to let people know about
this, it helps to have (an "In preparation" BibTeX entry).

With this commit, Pedram is added as an author of Warp and an empty BibTeX
entry has been added for Warp when users call it with '--cite'. Currently
the only stable thing about is its author list.