NEWS and mailmap: corrected Pedram's name to have a dash
Until now, Pedram's full name was written in the NEWS file as "Pedram
Ashofteh Ardakani". Recently, Pedram informed me that he would prefer to
have a dash in between the latter two components of his name to be easier
to recognize the first/last names.

With this commit, this has been corrected in the NEWS file and the
'.mailmap' file has also been updated to show his name in that format.