MakeProfiles: properly account for NaN pixels in custom inputs
Until now, when a custom input (profile or image) had a NaN pixel and the
user hadn't given the '--mcolnocustimg' or '--mcolnocustprof' options,
MakeProfiles would create a fully NaN image.

With this commit, the problem has been fixed in the step that we calculate
the sum of the raw profile: we now only add pixels to the sum that are not
NaN.

This bug was reported by Nafise Sedighi.

This fixes bug #63189.