Library (kdtree.h and match.h): can account for inputs with no data
Until now, 'gal_kdtree_create' always assumed that the input coordinates
dataset has atleast one element. Therefore, when it was given a dataset
with no rows (can happen in a pipeline!), it would create a segmentation
fault.

With this commit, 'gal_kdtree_create' now returns the NULL pointer in such
cases, and the 'gal_match_kdtree' function will also return NULL when the
input k-d tree is empty. Therefore when an empty table is given to the
Match program, it won't crash, it will just return a no-match output.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #63207.