MakeProfiles: ignore profiles where coordinates cannot be read
Until now, when a single RA/Dec coordinate couldn't be converted to image
coordinates for any reason (for example having a wrong value, or too far on
the sky), MakeProfiles would crash! Even if there were positions that could
be converted properly!

This was the result of the behavior until recently: even a single bad
conversion, would make all results blanks in the 'gal_wcs_world_to_img'
function. However, we now only set the values to NaN that WCSLIB flagged.

With this commit, using the new behavior of 'gal_wcs_world_to_img', we now
separte the rows that have a bad fit and ignore them in MakeProfiles.

This issue was reported by Sepideh Eskandarlou.