Book: copyright highlighted in the script of writing part of tutorial
Until now, the "Writing scripts to automate the steps" section of the
genenral tutorial didn't mention adding a copyright in a script. But it is
an essential part of any script that tells us who is writing the script,
under what license the script is made, etc.  Therefore, was is necessary to
introduce this concept (and how to do it) in this section for new script
writers.

With this commit, some explanations about adding a copyright statement in
the script have been added. Then in the last script, we added copyright
after the shebang to illustrate the most critical information that a script
should provide to readers. Also in this section, a link to the "Your
rights" and GNU GPL sections of the book have been added.