Library (warp.h): produce check image to visualize the Moire pattern
Until now, we could see Moire patterns on the warped images after aligning
with the WCS. But there was no way to quantify/visualize exactly where they
were most significant using a high-level option. The new '--checkmaxfrac'
serves this purpose.

With this commit, a new 'checkmaxfrac' component has been added to the
library's input/output structure. When it is non-zero, the library function
will add a 'next' component to the output 'gal_data_t' that will contain
the maximum fraction of a single input pixel in the output grid. added the
'--checkmaxfrac' option which appends an image to the final HDU/extension
(i.e. MAX-FRAC) that shows the maximum area covered on the input image
grid, by the created output image grid.

Also, to clarify the issue with the Moire pattern and how to deal with it,
a section called "Moire pattern" has been added in the Warp section of the
book.

Also these changes were done in this commit:

- Refactor: fix memory leak of the 'cdelt' dataset, previously it was
  replaced with NULL instead of 'gal_data_free'.

- Refactor: replace 'p->wa.*' with 'wa->*' in 'bin/warp/ui.c'.

- Refactor: use the 'fmax' GNU C Library 'double' comparison instead of the
  ternary operation. It is succing, more clear, and might have better
  performance.

- Book: reword where '--align' option was mentioned in the Gnuastro
  manual. This option was deprecated when the 'wcsalign' functions were
  added to the Warp library.

- Book: remind the readers that '--edgesampling' is not the same as
  oversampling!