Book: Deleted cd command in the script of Writing scripts tutorial
Until now, we used the 'cd' command in our commands to go to the directory
of the downloaded files and then return to the current directory. But there
is no need for the 'cd' command in the script and we can make a directory
and put our files in it from the running directory.

With this commit, the 'wget' command has been modified to use '-O' for
writing the downloaded file directly within the download directory (and not
need the 'cd' command). Also, the 'downloaddir' variable has been renamed
to 'dldir' to be shorter and help make the 'wget' comands fit in one line.