Book: tip for using multiple operators in one command of arithmetic
Until now, there was no explicit tip in the "Arithmetic operators" section
of the book on merging several consecutive Arithmetic commands into a
single command. While this is a natural consequence of the "Reverse polish
notation" (as described in its respective section), many readers who are
new to it may not notice this.

With this commit, this "tip" has been added as an item in a newly added
table of tips at the start of the "Arithmetic operators" section. The
previous box about NaN pixels in arithmetic has also been added as an item
there.

Two typos have also been fixed in other parts of the book.