Library (wcs.h): coverage func works with image that crosses RA=0
Until now, when the input image crossed the RA=0 hour circle, the outputs
of the 'gal_wcs_coverage' function were unreasonable (for example the width
for a ~1 degree image would be about 359 degres!).

With this commit, the function finds the dimension that may have this
problem (usually the RA dimension, but can also be the galactic or ecliptic
longitudes for example), checks if the width is larger than 180 degrees,
and if so, makes the necessary corrections to report the proper width and
minimum and maximum coordinates.

In the process, I also noticed an "may be used uninitialize" compiler
warning for the function pointer 'warp_pixel_perimeter' within the
'warp_pixelarea_onthread' function of 'lib/warp.c'. So it has been
initialized to NULL to have a clean build with no un-necessary warnings.

This bug was reported by Irene Pintos Castro.

This fixes bug #63257.