Book: fixed typos in the new Moire pattern section
Until now, the dithering section of the demo mistakenly used the same image
as Warp's input when preparing two fo the exposures. This defeated its
purpose: to show how dithering can reduce/remove Moire patterns in the
output image. Also, in one case, 'MAX-FRAC' was mistakenly written as
'FRAC-MAX'. Furthermore, I noticed that the new
'GAL_WARP_OUTPUT_NAME_WARPED' macro of 'lib/gnuastro/warp.h' was missed for
one of the dataset allocations in 'lib/warp.c' (which was still hardcoded
as "Aligned").

With this commit, these typos have been fixed in the source and book and
some parts of the "Moire pattern" section have been edited to be more
clear (including its name).