Library (txt.h): precision of 0 in floats, prints rounded integer
Until now, when a floating point column (32-bit or 64-bit) was printed with
a precision of 0 (for example using '--txtf64precision'), the option would
be ignored! This was happening because in the text library, we were
assuming that the precision is always larger than 0 (so when it was given a
value of zero, it would use the default value instead).

With this commit, 0 is no longer the invalid value, but we are properly
using the 'GAL_BLANK_INT' value as an invalid value and this fixed the
problem.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #63266.