Library (error.h): Added more functions
Until now, the error library only had print and allocate functions.

With this commit, we add the is_warning flags to the `gal_error_t`
structure. This flag indicates whether an error is just a warning or
a breaking error(indicating that the program should be stopped).
This severity of each error is determined by the error code defined in
each module. The value assigned to each error code inside a module's
include file, indicates its `is_warning` flag status.