Library(cosmology.c): Error Handling
Until now, the error codes for testing the new error handling methods
in cosmology library were random. This meant, that the `is_warning`
flags were also just set to their default values i.e 0.

With this commit, we define error codes for the different types of errors
in the cosmology library. We also define a stringizing function. Thus, the
error codes will be string constant passed stringizing the macros, and the
`is_warning` flag of each error will be the actual value of that macro.