Fits: --list* options now print HDU names in the input's order
Until now, the '--listimagehdus' or '--listtablehdus' options of the Fits
program would print the HDUs in reverse order! This was because we were
simply writing a last-in-first-out list (most efficient!) and finally
printing it. But this is not what the user would expect.

With this commit, the list is inversed after it has been found. This fits
better with what a user would expect.

This was noticed after a discussion with Michael Stein and Sepideh
Eskandarlou.