Table: short format of --txtf64format to -A (due to conflict)
Until now, the short format of the '--txtf64format' option was
'-d'. However, this was the same as the short format for the '--descending'
option and this caused a conflict when '--decensing' was used.

With this commit, the short format '-A' has been set for '--txtf64format'
and this fixed the problem.

This bug was reproted by Giulia Golini.

This fixes bug #63270.