Crop: new option to let users set the name of the output HDU
Until now, if the user wanted to set the name ('EXTNAME' keyword) of the
output HDU, they would have to call 'astfits' afterwards. But this is a
pretty basic operation that like Arithmetic, the user expects to do in the
same command that they produced the HDU.

With this commit, based on the experience with Arithmetic, a new
'--metaname' option has been added to Crop (same long version name as the
same option in Arithmetic).