Crop: new --append operator to add crop as new HDU fo output
Until now, Crop would always delete the output file if it already
existed. However, in some scenarios, it is necessary keep existing HDUs and
just append the new crop to them. To do this, it was necessary to put the
cropped HDU in a temporary file, then use 'astfits' to copy that HDU into
the file with existing HDUs. But this is an annoying extra step (given that
the low-level structure to preserve the HDUs is already present in crop).

With this commit, a new '--append' option has been added for this job. When
it is called, if the output file already exists, the cropped HDU will be
written after the last HDU. An example usage of this to crop all image-HDUs
of a file has been added at the example commands in the "Invoking Crop"
section of the book.

In the process, I also noticed that Crop's output options weren't listed in
the "Crop output" sub-section of the "invoking Crop" section. They have
also been moved there with this commit.

This issue was implemented after a discussion with Michael Stein and Elham
Saremi.