Book: corrected two minor typos in newly added example
Until now, the example in the book for cropping multiple HDUs of a FITS
file had two typos: the 'hdu' variable was read as '$$hdu' and there was no
'--width'. Both would cause a crash.

With this commit, a single '$' has been used for the 'hdu' variable and a
width has been added.