Library (type.c): gal_type_to_string uses 6 and 14 digits for floats
Until now, this function (which is used to print numbers as strings) was
using '%.6g' and '%.10g' for 32-bit and 64-bit floating point
numbers. However, the '%g' format will (in some situations) not print
statistically significant digits and generally, it is hard to predict if it
will print a fixed point or exponential notation.

With this commit, the printing formats for this function have been changed
to '%.6f' and '%.14f'. This will always print fixed point, and will print
the required number of digits after the decimal point. Making the result
more predictable and without any possible loss of precision.