Book: minor edits in the extended PSF section and fixed some typos
Until now, to start the discussion on "astscript-psf-stamp" in the
tutorial, there was no introduction. We had just mentioned that we now use
this script.

With this commit, a short introduction for this script is added and some
references in the first of the "Building outer part of the PSF" section
have been corrected.  Also, some minor typos are fixed.