NoiseChisel: no. tiles for outliers given through --outliernumngb
Until now, the number of tiles used for identifying outlier tiles (on the
wings of bright galaxies) was given through the '--interpnumngb'
option. But as the name suggests, '--interpnumngb' is for the number of
tiles to use for the interpolation phase. Using the same value for outlier
rejection was both confusing and made it hard to customize in some
situations (since interpolation is done after outlier rejection).

With this commit, we now have a new option called '--outliernumngb' that is
only for identifying the number of tiles for finding outliers. The number
of neighboring tiles for interpolation is now uniquely for interpolation.

Therefore '--outliernumngb' is now the most imporant option to change when
the outer wings of bright galaxies/stars are not detected.

This was done after a discussion with Elham Saremi.