Library (error, cosmology): Added bit ordered macros
Until now, the value of error codes(macros) defined in `cosmology.h` only
indicated the `is_warning` flag status, and the error `code` was a string
constant obtained by stringizing the macro name. However, this approach
leads to more work on the user side while detecting errors and also has
grater storage requirements owing to string error codes.

With this commit, we change the error `code` to be a 16-bit unsigned int.
The errors codes will be globally numbered from 1. The 32 bit macro values
defined in headers will be set such that their least significant 16-bits
represent `is_warning` and msot significant 16-bits represent the `code`.

To allow this, we add a `BITSET` macro function in the library headers that
returns the error `macro_val` given the `code` and `is_warning` flag.

Also, added a error check function in error library to allow users to check
if a particular type of error exists in an error structure.