Cosmiccal: Added arguments required by error lib
Until now, for the error/sanity checks for the binary programs
were performed in their `ui.c` files. Before passing the values
to the actual code of the program.

With the addition of the `error.c` library, we can fortify this system
further by adding verfication checks in the main programs too,
incase any errors were missed by the ui sanity checks.

With this commit, we add the additional error argument required by
the cosmology programs to the library calls made in `cosmiccal.c`.