Library(error.h): Added Library codes to error
Until now, the `gal_error_t` structure only had the `code` and
`is_warning` flags as attributes. However, this wouldve resulted in having
to create a global code system for all the errors in the entire library.

With this commit, we avoid this problem by simple adding a 8 bit integer
`lib_code` as another attribute. This indicates the library which generated
the error. The lib_codes are stored as an enum where each module of
the library is written in alphabetical order.

So now, the 32 bit error macro values, will have the least significant 8
bits represent the `is_warning` flag, the next significant 8 bits represent
the `code` and the next 8 bits represent the `lib_code`.

Also made changes to all functions in the error library to facilitate this
new attribute.